Remove the QKnxByteStoreRef as class member
The reference is no longer a member of byte store, we will create it
on demand. This enables the option of passing an index as argument to
point into the data vector the byte store holds. Couple the ref class
and the store more thightly, so the store can only create the ref,
while only the ref can access the stores raw data.

Change-Id: I7b6a375cee04f8e0738f343f2e2e41528939ac17