Initial implementation of KNX a DPT factory and some DPT's, auto-test
The idea is to create DPT's with a common base class and register
them in a factory. Since we cannot implement all of the available
DPTs, the factory offers functionality to register user implemented
types or even override implemented ones. Access to a given DPT is
realized by giving the main type and according sub type.

Change-Id: Ifc4d0e16503d3d250dad310b2c269d5e73a03573
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>