Knx UI Demo interacting with ETS experimental board
The demo shows the capability of the QtKnx module to communicate with
a KNX system. The qmlknxdemo class holds the state of the lights on
the board and serves as an interface towards the QtKnx library. The
checkboxes ids found in the KnxBoardLayout.qml use the same naming
convention as the ones used by the ETS board.

An IP tunnel between the demo application and a KNXnet/IP router is
established. Within this tunnel the whole KNX protocol is
encapsulated. This allows the application to control the light status
and monitorize changes done from the ETS board.

Change-Id: I69db3c9fa2af59778a32f8e3fa03165dd6651881
Reviewed-by: Andrew O'Doherty <andrew.odoherty@qt.io>
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>