Rewrite the KNXnet/IP frame class to be generic
Introduce builder and proxy classes to deal with specializations
of the KNXnet/Ip frame class. Adapt existing code and auto-tests.

For the moment the new class has an QKnxNetIpFrame{Ex} suffix, so
that already existing code can be adapted. Next commit will remove
the 'old' QKnxNetIpFrame class and remove the {Ex} suffix.

Change-Id: I06ba590e09f005a9e778ddf9b53e8a21c87ec880
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>