Fix routing busy wait time setter implementation
According to 03_08_05 Routing (2.3.5 Flow control handling):

If the incoming queue of a KNXnet/IP Router or KNX IP device exceeds
the number of Datagrams that can be processed (e.g. sent to the local
KNX Subnetwork by a KNXnet/IP Router) within a period of T(process)
then this device SHALL send a ROUTING_BUSY Frame with a wait time tw.
The default value for T(process) SHOULD be 100 ms and MAY be greater
than 100 ms.

Change-Id: Ic387fd2f78372563aa17eafa043ff963dc192230
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>