Avoid sending a multicast search request via loopback device
Initially the QKnxNetIpServerDiscoveryAgent was joining the
multicast group on the default interface chosen by the operating
system. This lead to strange behavior on windows with several
interfaces up and running. A fix for this was applied in change
4d5a04a0793ef05ade54a09cd469c9934f102753 that explicitly defines
the interface to use. No if we use localhost as default address,
the loopback device gets picked and no multicast search request
will be send to the network and one would always need to specify
the target address to respond to, making multicast obsolete.

Change-Id: I4fd0fbff72d0a6f7b3e497976d70293aa8decc64
Reviewed-by: Andrew O'Doherty <andrew.odoherty@qt.io>
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>